+++ src/importexport/braille/internal/exportbraille.cpp
@@ -1475,7 +1475,7 @@ QString ExportBrailleImpl::brailleChordInterval(Note* rootNote, const std::vecto
     if (interval == 1 && rootNote->octave() == note->octave()) {
         noteOctaveBraille = brailleOctave(note->octave());
     }
-    int noteIdx = mu::indexOf(notes, note);
+    size_t noteIdx = mu::indexOf(notes, note);
     int intervalWithPreviousNoteInChord = computeInterval(notes.at(noteIdx - 1), note, false);
     // (b) it is the first or only interval and is more than an octave from the written note,
     if (noteIdx == 1 && intervalWithPreviousNoteInChord > 8) {
